[ API ] User Endpoint Modularization
